# openApi-yaml.yaml
openapi: 3.0.1
info:
  title: CMS
  description: REST APIs for Contact Management System
  version: "2.0"
servers:
- url: http://localhost:8090
  description: Generated server url
paths:
  /cms/v1/createOrUpdateContact:
    post:
      tags:
      - cms-controller
      operationId: createContact
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/cmsRequest'
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/BaseApiResponse'
  /cms/v1/readContact:
    get:
      tags:
      - cms-controller
      operationId: viewContact
      parameters:
      - name: firstName
        in: query
        required: false
        schema:
          type: string
      - name: lastName
        in: query
        required: false
        schema:
          type: string
      - name: email
        in: query
        required: false
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/BaseApiResponse'
  /cms/v1/deleteContact:
    delete:
      tags:
      - cms-controller
      operationId: deleteContact
      parameters:
      - name: id
        in: query
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/BaseApiResponse'
...
components:
  schemas:
    cmsRequest:
      type: object
      properties:
        id:
          type: integer
          format: int64
        firstname:
          type: string
        lastName:
          type: string
        email:
          type: string
        phoneNumber:
          type: string
    BaseApiResponse:
      type: object
      properties:
        responseStatus:
          $ref: '#/components/schemas/ResponseStatus'
        responseData:
          type: object
        message:
          type: string
        statusCode:
          type: integer
          format: int32
          writeOnly: true
    ResponseStatus:
      type: object
      properties:
        statusCode:
          type: integer
          format: int32
...