# src/main/java/com/cms/request/cmsRequest.java
package com.cms.request;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

@Builder
@Data
@AllArgsConstructor
@NoArgsConstructor
public class cmsRequest {

	private Long id;
	private String firstname;
	private String lastName;
	private String email;
	private String phoneNumber;
}